rename document table to paper
i think it's going to cause me a headache to do this later and this
should make it easier to manage other data types we want to add for
processing. we might create a base document type that the paper, and
others in the future, could inherit from. for now, this is sufficient.