start backend changes for honoring subscription tier
two main settings are being honored first: number of paper uploads and
knowledge base size.

to calculate the size of the knowledge base, we want to use the
`size_in_kb` property pulled from each paper. we haven't been
calculating this historically, so there will be some `None` values. to
accommodate, calculate on the fly using the s3 API.

todo - we still have to add logic that calculates the size of the file
to upload and checks against the knowledge base size to verify that it's
allowed. for now, it's just checking that it's not already over.