get accurate calculations for relevant subscription usage details
use different crud methods for calculation number of paper uploads,
knowledge base size, weekly audio and chat usage.

i think we'll switching the pricing usage for chat and audio to weekly
to make it easier to track, and to make the jitter of daily changes less
disruptive. hopefully the usage limits will still be generous enough
that value can be extracted from the app sufficiently.