add support for server-side searches that include highlights and annotations
for better cross-knowledge base search, look over paper text,
highlights, annotations against the search query.

limitations: this does not support any fuzzy search at the moment. we
can extend this later on to include better fuzzy / approximate search,
potentially just through a good regular expression, if not
embeddings-assisted search.

there are primarily two issues with embeddings-supported search. 1. it
is troublesome to maintain the database as migrations are required to
change models and create the index. and 2. it provides limited
observability into why a particular result produced a match.

a tertiary issue is that search quality relies on a robust chunk
strategy, to some extend.