for initial PDF processing, let the backend take care of uploading the file to s3
we're seeing that there's a significant delay in the initial upload to
s3 taking place when submitting the job to the message broker. this is
happening with a direct relationship to file size, indicating that it's
likely due to the size of the packet being stored in MQ, especially when
the file exceeds 3,4 MB. eat this latency before upload, sending the
file to s3 and passing the object key to the celery service.

the celery service then downloads the file from s3 and processes it as
usual. this is subtoptimal because we are now eating 2 roundtrip costs
to s3, rather than just the 1 we were earlier, but the issue is that
services like rabbit mq aren't built for transferring large data
packets.