completely untested as of yet, start working on the chat with everything endpoint
this needs to work by collecting all papers and sending the research
agent on an evaluation journey to find relevant data and evaluate
whether it can find knowledge to sufficiently answer the user's
question.

to facilitate these behaviors, we need to migrate the database in order
to support conversatoins that can be associated without papers, using
different types. this will later also be extended to support handling of
the project chats, and possibly also chats in the annotations!

the critical concept behind chats with the entire knowledge base is that
we need tool calling that sufficiently enables the AI to leverage the
knowledge base to answer full corpus questions.

some initial wiring and data migrations have taken place, but they are
untested end-to-end.