add crud and api methods for managing projects
projects can have chats, papers, and audio overviews. create associated
crud and api messages for managing projects. create a project crud base
class. we extend from the existing base class to create another one,
because authentication hinges on the `projectrole` association table
lookups. this is because we'll have multiple people who can be members
of each project. multiplayer features will be implemented later. to
start with, we'll make an assumption that the paper owner is the prject
owner, and adapt this assumption as team features are added in.