start adding in project views
start adding in the scaffolding for what will be the projects dashboard,
project management view. begin setting up the features for uploading and
attahing papers to projects. the UI is still being worked out.

next step, we'll add functionality to start chats in projects, and then
create audio overviews eventually.