extend the multipaper chat operations to support projects
projects and everything chat are quite similar in nature, except that
the base for how eligible papers are determine in the file_tools is
different. we have to be careful with the file selection process here,
as the quality and security of the app depends on correct selection
critiera.

start work on integration to the chat and conversation APIs as well