inital changes to separate out the conversations view
the understand page and the projects page have a _mostly_ similar layout
for conversations. user message, ai response, multi-paper citations. as
such, re-use the UI as much as possible to minimize code duplication and
ease maintenance down the road. the less UIs we have to keep updating,
the better.

add a basic conversations projects page to start with that just fetches
hstory and renders. haven't tested out the integration with backend APIs
yet.