the final response API call to the LLM in multi-paper chat can be really long
the final response APi call to the LLM in muilti-paper chat can be
really long, like really really long. in order to improve the
reliability of our server-client connection TTL, use an asyncio wrapper
to send periodic pings to the client to let it know it's alive.

this is a short term workaround. it let's us buy some time to make to do
with the long query response times until we can improve latency and
hopefully add more resilience to the longer-running responses being
picked back up by the client when disconnected.