handle more edge cases in the sign in APIs
for google sign in, if the user already exists with a different
provider, error out.
send an additioanl parameter, newly_created, for the email flow in order
to dynamically adapt the sign-in with email logic.
add another API for explicitly setting the name of the user in the email
flow, since we don't get an authenticated name from an oauth provider
here.
fix the response of the verify endpoint to send a json object with a
redirect parameter, rather than directly sending a redirect response,
which was encountering some CORS issues.