change the PaperCard in case the user does not own the paper
if the user doesn't ownt he paper in the project that they clicked on,
then show a preview of the paper, instead of the full paper reader view.

eventually, hook up the duplicate button to give them the option to add
it to their own library.

if they've duplicated, then show the duplicated paper via one-click in
their own library.