perf improvements for project agent chat
improve the speed of the project agent chat by migration the execution
of the search query to use a database-only approach (with postgres full
text search and phrase vectors). this avoids the expensive in-memory
python processing of the full texts of the files. use this in the
search_all_files tool execution.