update the backend API for fetching all papers for the user to incl. all 500
issue: we were hard-coding the return limit to 100. this meant that any
users who have >100 uploaded papers hit an issue where older papers
weren't being returned. fix this so that all the papers are included in
the response. in order to do so, add a step that quickly fetches updated
s3 presigned urls for the files.