use tanstack to virtualization rendering of all papers
this is helpful to render the large number of documents that will be
returned by the updated backend papers API. we want to return all 500 of
the documents. in order to render these without killing the browser, use
virutalization to render only the displayed + a few additoinal at a
time.

this means that the search/filtering should still work in browser
memory, as it was before. this keeps it lightweight and quick.