track whether metadata extraction has already been attempted
for latency's sake, we don't want to continually re-attempt metadata
extraction for PDFs that failed extraction
on the other hand, the failure could have been transient, or the data
may not be available yet. so, allow a re-attempt.
the final issue is that some papers may just not be academic in nature,
so we would continually be polling irrelevant data. perhaps there should
be a maximal time interval for these attempts.