make the initial summary more minimalist
go with a more minimalistic initial summary. set this as the first
message in the paper chat.
this in theory should help reduce the indexing times and make the
initial read of the paper more approachable.
this helps the user drill down deeper into whatever part of the paper is
more relevant to them.
render this as the first message, and don't show the overview page if
the summary is not present.
also, get rid of the paper metadata section.