add support for properly using function call responses
LLM API providers now offer a first-party paradigm for passing along the
results of a tool call session to the API - via tool call results. this
helps the multi-turn tool-use agent better determine which tools to use
in subsequent runs.

prior, we were manually constructing the previously collected evidence
from the prev function calls and passing it into the user prompt. this
is improved because it leverages whatever RL training the models have
undergone to know that these are results from previous executions.